+++ websocket/server.go
@@ -1,6 +1,3 @@
-// Copyright 2013 The Gorilla WebSocket Authors. All rights reserved.
-// Copyright 2016 Gerasimos Maropoulos. All rights reserved.
-
 package websocket
 
 import (